requirements.txt
simplematrixbotlib>=2.8.1
PyYAML>=6.0
